bug fixing (rx symbol error)
