+++ code/spikes/publish.001/src/ui/Root.HoldingPattern/entry.tsx
@@ -1,3 +1,5 @@
+import 'symbol-observable';
+
 import { createRoot } from 'react-dom/client';
 import { RootHolding } from '.';
